Updated README Added :time type to formation module Changed exception type from MotionModel::PersistFileFailureError to MotionModel::PersistFileError when persistence operations fail.

+++ lib/motion_model/model/model.rb
@@ -191,7 +191,7 @@ module MotionModel
       def destroy_all
         ids = self.all.map{|item| item.id}
         bulk_update do
-          ids.each do |item| 
+          ids.each do |item|
             find(item).destroy
           end
         end
@@ -439,7 +439,7 @@ module MotionModel
     def delete
       call_hooks('delete') do
         target_index = collection.index{|item| item.id == self.id}
-        collection.delete_at(target_index)
+        collection.delete_at(target_index) unless target_index.nil?
         self.class.issue_notification(self, :action => 'delete')
       end
     end
